+++ LoggingService/LoggingService.cs
@@ -4,14 +4,6 @@ using static Hypertherm.Logging.LoggingService;
 
 namespace Hypertherm.Logging
 {
-    public interface ILoggingService
-    {
-        void ClearLog();
-        void DumpLog();
-        bool isError();
-        void Log(string message, MessageType type);
-    }
-
     public class LoggingService : ILoggingService
     {
         private string _filename;
