+++ OidcAuthService/IAuthenticationService.cs
@@ -1,0 +1,11 @@
+
+using System.Threading.Tasks;
+
+namespace Hypertherm.OidcAuth
+{
+    public interface IAuthenticationService
+    {
+        Task<string> Login(string user = "default-user");
+        void Logout(string user = "default-user");
+    }
+}
