+++ cc-cli/cc-cli.csproj
@@ -16,7 +16,10 @@
     <None Update="authconfig.json" CopyToOutputDirectory="Always" />
   </ItemGroup>
   <ItemGroup>
+    <PackageReference Include="Microsoft.Extensions.Http" Version="3.1.5" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
+    <PackageReference Include="Polly" Version="7.2.1" />
+    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="2.1.0" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\AnalyticsService\AnalyticsService.csproj" />
